apps/api/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic==2.11.7
pydantic-settings==2.10.1
sqlalchemy==2.0.43
psycopg[binary]==3.2.10
alembic==1.16.5
redis==6.4.0
celery==5.5.3
httpx==0.28.1
python-multipart==0.0.20
passlib[argon2]==1.7.4
pyjwt==2.10.1

pytest==8.4.1
pytest-asyncio==1.1.0
ruff==0.12.10
mypy==1.18.2
types-passlib
email-validator